Todo 7: Create two classes for Freshman and Senior ToDo 8: The senior class should have a minimum of 85 credits

+++ src/main/java/com/mycompany/csc325_oop_designreview_lab/MainClass.java
@@ -11,17 +11,12 @@ package com.mycompany.csc325_oop_designreview_lab;
 public class MainClass {
 	
 	public static void main(String[] args) {
-
-                
-                // Todo 7: Create two classes for Freshman and Senior 
-                
-                // ToDo 8: The senior class should have a minimum of 85 credits  
-		
+	
 		// ToDo 9: Add a toString method for Freshman class
 		
                 // ToDo 10: Add a toString method for Senior class
              
-		Freshman std1= new Student("James", (short)20, 12); // name, age, credits
+		Freshman std1= new Student("James", 20, 12); // name, age, credits
                 
                 Senior std2 = new Student("John", 30, 90);
